refactor : "binding 구조 정리"
auth_binding 분리, navigation_binding 공통 의존성만 유지
post_binding PostViewController tag 기반 등록으로 변경
post_detail_view_binding 제거

Resolve: #176
