+++ lib/core/router/bindings/post_binding.dart
@@ -12,35 +12,38 @@ import '../../../presentation/post/controllers/post_view_controller.dart';
 
 class PostBinding extends Bindings {
   final int postId;
+
   PostBinding(this.postId);
 
   @override
   void dependencies() {
     Get.lazyPut<PostRemoteDatasource>(
-          () => PostRemoteDatasourceImpl(Get.find<AuthClient>()),
+      () => PostRemoteDatasourceImpl(Get.find<AuthClient>()),
     );
     Get.lazyPut<PostRepositoryImpl>(
-          () => PostRepositoryImpl(Get.find<PostRemoteDatasource>()),
+      () => PostRepositoryImpl(Get.find<PostRemoteDatasource>()),
     );
     Get.lazyPut<GetPostDetailUseCase>(
-          () => GetPostDetailUseCase(Get.find<PostRepositoryImpl>()),
+      () => GetPostDetailUseCase(Get.find<PostRepositoryImpl>()),
     );
     Get.lazyPut<CreatePostUseCase>(
-          () => CreatePostUseCase(Get.find<PostRepositoryImpl>()),
+      () => CreatePostUseCase(Get.find<PostRepositoryImpl>()),
     );
     Get.lazyPut<UpdatePostUseCase>(
-          () => UpdatePostUseCase(Get.find<PostRepositoryImpl>()),
+      () => UpdatePostUseCase(Get.find<PostRepositoryImpl>()),
     );
     Get.lazyPut<DeletePostUseCase>(
-          () => DeletePostUseCase(Get.find<PostRepositoryImpl>()),
+      () => DeletePostUseCase(Get.find<PostRepositoryImpl>()),
     );
     Get.lazyPut<LikePostUseCase>(
-          () => LikePostUseCase(Get.find<PostRepositoryImpl>()),
+      () => LikePostUseCase(Get.find<PostRepositoryImpl>()),
     );
     Get.lazyPut<UnlikePostUseCase>(
-          () => UnlikePostUseCase(Get.find<PostRepositoryImpl>()),
+      () => UnlikePostUseCase(Get.find<PostRepositoryImpl>()),
     );
+    //TODO : tag가 메모리에 남는 경우 delete 로직 추가
     Get.put<PostViewController>(
+      tag: postId.toString(),
       PostViewController(
         postId: postId,
         useCase: Get.find<GetPostDetailUseCase>(),
